Add support for multiple linked worktrees
Until now, the working tree was forced to be the parent of a found
GIT_DIR. This is true for plain repositories, but fails for worktrees
and submodules. The GIT_DIR for those special repositories resides
inside the main now.

This is not supported by JGit out-of-the-box as of now. So we try to
find the `commondir` of a worktree and set it as the repository‘s
GIT_DIR.

See https://git-scm.com/docs/gitrepository-layout for further
information.

See #1
Closes #98